+++ app/Helpers/Parameters.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace App\Helpers;
 
-use Exception;
-
 class Parameters {
 	/**
 	 * @var array
@@ -21,5 +19,12 @@ class Parameters {
 	 */
 	public function getSiteTitle(string $locale): ?string {
 		return $this->parameters['siteTitle'][$locale] ?? null;
+	}
+
+	/**
+	 * Get the short event name for the given locale, if defined.
+	 */
+	public function getSiteTitleShort(string $locale): ?string {
+		return $this->parameters['siteTitleShort'][$locale] ?? null;
 	}
 }